+++ nautilus_trader/adapters/bybit/schemas/ws.py
@@ -212,18 +212,19 @@ class BybitWsOrderbookDepth(msgspec.Struct):
             for d in self.a
         ]
 
-        num_bids_raw = len(bids_raw)
-        num_asks_raw = len(asks_raw)
         deltas: list[OrderBookDelta] = []
 
         if snapshot:
             deltas.append(OrderBookDelta.clear(instrument_id, 0, ts_event, ts_init))
 
-        for bid_id, bid in enumerate(bids_raw):
+        bids_len = len(bids_raw)
+        asks_len = len(asks_raw)
+
+        for idx, bid in enumerate(bids_raw):
             flags = 0
-            if bid_id == num_bids_raw - 1 and num_asks_raw == 0:
+            if idx == bids_len - 1 and asks_len == 0:
                 # F_LAST, 1 << 7
-                # Last message in the packet from the venue for a given `instrument_id`
+                # Last message in the book event or packet from the venue for a given `instrument_id`
                 flags = RecordFlag.F_LAST
 
             delta = parse_bybit_delta(
@@ -239,11 +240,11 @@ class BybitWsOrderbookDepth(msgspec.Struct):
             )
             deltas.append(delta)
 
-        for ask_id, ask in enumerate(asks_raw):
+        for idx, ask in enumerate(asks_raw):
             flags = 0
-            if ask_id == num_asks_raw - 1:
+            if idx == asks_len - 1:
                 # F_LAST, 1 << 7
-                # Last message in the packet from the venue for a given `instrument_id`
+                # Last message in the book event or packet from the venue for a given `instrument_id`
                 flags = RecordFlag.F_LAST
 
             delta = parse_bybit_delta(